Move fn to common location
